chore: deprecate core module
Remove the catch-all core module by promoting the implementation to the
top level kyoto module. Further refactors can break this up into existing
and new feature specific modules.
