+++ src/network/peer_map.rs
@@ -21,23 +21,19 @@ use tokio::{
 
 use crate::{
     chain::HeightMonitor,
+    channel_messages::{CombinedAddr, MainThreadMessage, PeerThreadMessage},
     db::{traits::PeerStore, PeerStatus, PersistedPeer},
+    dialog::Dialog,
+    error::PeerManagerError,
     network::{
         dns::DnsResolver,
         error::PeerError,
         peer::Peer,
         traits::{ClearNetConnection, NetworkConnector},
+        PeerId, PeerTimeoutConfig,
     },
     prelude::{default_port_from_network, Median, Netgroup},
-    ConnectionType, PeerStoreSizeConfig, TrustedPeer,
-};
-
-use super::{
-    channel_messages::{CombinedAddr, MainThreadMessage, PeerThreadMessage},
-    dialog::Dialog,
-    error::PeerManagerError,
-    messages::Warning,
-    PeerId, PeerTimeoutConfig,
+    ConnectionType, PeerStoreSizeConfig, TrustedPeer, Warning,
 };
 
 const MAX_TRIES: usize = 50;
